% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fvarug-package.r
\docType{package}
\name{fvarug-package}
\alias{fvarug-package}
\title{FVA R User Group}
\description{
Codes and functions written for or in meetings of the FVA R User Group.
}
\note{
this is a collection of potentially highly volatile codes and
functions used as examples at the R User Group at the Forstliche Versuchs- und
Forschungsanstalt Baden-Wuerttemberg.
}
\section{Version}{
 $Id: 21e3a57157dcd729e5bbed73ed5d84159e8ea554 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
\keyword{package}

