% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library_calls.r
\name{get_package_list}
\alias{get_package_list}
\title{List packages}
\usage{
get_package_list(path = NULL, pattern = "*\\\\.[rR]$")
}
\arguments{
\item{path}{The path to look for source files. If NULL, the user is queried
to give a path.}

\item{pattern}{A pattern defining source files.}
}
\value{
a named list containing
\itemize{
\item "path", a character vector giving the path,
\item "pattern", a character vector giving  the pattern,
\item "packages_installed", a character vector giving the packages installed,
\item "package_calls", a factor giving the package calls.
}
}
\description{
List packages installed on your system and called from source files under a
given path.
}
\section{Version}{
 $Id: 9bbb752b06d887f2115e37c3e9dadd89e40c49c7 $
}
\examples{
get_package_list()
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

