% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library_calls.r
\name{plot.package_list}
\alias{plot.package_list}
\title{Plot function for objects of class "package_list"}
\usage{
\method{plot}{package_list}(x, ...)
}
\arguments{
\item{x}{an object of class package_list.}

\item{...}{argument passing for generic \code{\link{plot}} compablility.}
}
\value{
invisibly NULL
}
\description{
Plot the frequencies of packages called.
These frequencies are given by the
\code{\link{get_package_list}}()[["package_calls"]].
}
\section{Version}{
 $Id: 9bbb752b06d887f2115e37c3e9dadd89e40c49c7 $
}
\examples{
plot(get_package_list())
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

