% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library_calls.r
\name{summary.package_list}
\alias{summary}
\alias{summary.package_list}
\title{Summary function for objects of class "package_list"}
\usage{
\method{summary}{package_list}(object, ...)
}
\arguments{
\item{object}{an object of class object.}

\item{...}{argument passing for generic \code{\link{summary}} compablility.}
}
\value{
invisibly NULL
}
\description{
Summarize the findings of \code{\link{get_package_list}}.
}
\section{Version}{
 $Id: 9bbb752b06d887f2115e37c3e9dadd89e40c49c7 $
}
\examples{
\dontrun{summary(get_package_list())}
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

